+++ py-polars/tests/unit/datatypes/test_struct.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 import typing
 from dataclasses import dataclass
-from datetime import date, datetime, time
+from datetime import datetime, time
 
 import pandas as pd
 import pyarrow as pa
@@ -216,16 +216,6 @@ def test_nested_struct() -> None:
     assert isinstance(nest_l2.dtypes[0], pl.datatypes.Struct)
     assert [f.dtype for f in nest_l2.dtypes[0].fields] == nest_l1.dtypes
     assert isinstance(nest_l1.dtypes[0], pl.datatypes.Struct)
-
-
-def test_eager_struct() -> None:
-    with pytest.raises(pl.DuplicateError, match="multiple fields with name '' found"):
-        s = pl.struct([pl.Series([1, 2, 3]), pl.Series(["a", "b", "c"])], eager=True)
-
-    s = pl.struct(
-        [pl.Series("a", [1, 2, 3]), pl.Series("b", ["a", "b", "c"])], eager=True
-    )
-    assert s.dtype == pl.Struct
 
 
 def test_struct_to_pandas() -> None:
@@ -685,55 +675,10 @@ def test_struct_getitem() -> None:
     ).to_dict(False) == {"a": [1]}
 
 
-def test_struct_broadcasting() -> None:
-    df = pl.DataFrame(
-        {
-            "col1": [1, 2],
-            "col2": [10, 20],
-        }
-    )
-
-    assert (
-        df.select(
-            pl.struct(
-                [
-                    pl.lit("a").alias("a"),
-                    pl.col("col1").alias("col1"),
-                ]
-            ).alias("my_struct")
-        )
-    ).to_dict(False) == {"my_struct": [{"a": "a", "col1": 1}, {"a": "a", "col1": 2}]}
-
-
 def test_struct_supertype() -> None:
     assert pl.from_dicts(
         [{"vehicle": {"auto": "car"}}, {"vehicle": {"auto": None}}]
     ).to_dict(False) == {"vehicle": [{"auto": "car"}, {"auto": None}]}
-
-
-def test_suffix_in_struct_creation() -> None:
-    assert (
-        pl.DataFrame(
-            {
-                "a": [1, 2],
-                "b": [3, 4],
-                "c": [5, 6],
-            }
-        ).select(pl.struct(pl.col(["a", "c"]).suffix("_foo")).alias("bar"))
-    ).unnest("bar").to_dict(False) == {"a_foo": [1, 2], "c_foo": [5, 6]}
-
-
-def test_concat_list_reverse_struct_fields() -> None:
-    df = pl.DataFrame({"nums": [1, 2, 3, 4], "letters": ["a", "b", "c", "d"]}).select(
-        [
-            pl.col("nums"),
-            pl.struct(["letters", "nums"]).alias("combo"),
-            pl.struct(["nums", "letters"]).alias("reverse_combo"),
-        ]
-    )
-    result1 = df.select(pl.concat_list(["combo", "reverse_combo"]))
-    result2 = df.select(pl.concat_list(["combo", "combo"]))
-    assert_frame_equal(result1, result2)
 
 
 def test_struct_any_value_get_after_append() -> None:
@@ -867,28 +812,6 @@ def test_sort_structs() -> None:
     }
 
 
-def test_struct_args_kwargs() -> None:
-    df = pl.DataFrame({"a": [1, 2], "b": [3, 4], "c": ["a", "b"]})
-
-    # Single input
-    result = df.select(r=pl.struct((pl.col("a") + pl.col("b")).alias("p")))
-    expected = pl.DataFrame({"r": [{"p": 4}, {"p": 6}]})
-    assert_frame_equal(result, expected)
-
-    # List input
-    result = df.select(r=pl.struct([pl.col("a").alias("p"), pl.col("b").alias("q")]))
-    expected = pl.DataFrame({"r": [{"p": 1, "q": 3}, {"p": 2, "q": 4}]})
-    assert_frame_equal(result, expected)
-
-    # Positional input
-    result = df.select(r=pl.struct(pl.col("a").alias("p"), pl.col("b").alias("q")))
-    assert_frame_equal(result, expected)
-
-    # Keyword input
-    result = df.select(r=pl.struct(p="a", q="b"))
-    assert_frame_equal(result, expected)
-
-
 def test_struct_applies_as_map() -> None:
     df = pl.DataFrame({"id": [1, 1, 2], "x": ["a", "b", "c"], "y": ["d", "e", "f"]})
 
@@ -899,24 +822,6 @@ def test_struct_applies_as_map() -> None:
     ).to_dict(False) == {
         "x": [{"x": "a", "y": "dd"}, {"x": "b", "y": "ee"}, {"x": "c", "y": "ff"}]
     }
-
-
-def test_struct_with_lit() -> None:
-    expr = pl.struct([pl.col("a"), pl.lit(1).alias("b")])
-
-    assert (
-        pl.DataFrame({"a": pl.Series([], dtype=pl.Int64)}).select(expr).to_dict(False)
-    ) == {"a": []}
-
-    assert (
-        pl.DataFrame({"a": pl.Series([1], dtype=pl.Int64)}).select(expr).to_dict(False)
-    ) == {"a": [{"a": 1, "b": 1}]}
-
-    assert (
-        pl.DataFrame({"a": pl.Series([1, 2], dtype=pl.Int64)})
-        .select(expr)
-        .to_dict(False)
-    ) == {"a": [{"a": 1, "b": 1}, {"a": 2, "b": 1}]}
 
 
 def test_struct_unique_df() -> None:
@@ -952,15 +857,6 @@ def test_nested_struct_logicals() -> None:
     # double nested
     payload = [[[{"a": time(10)}]], [[{"a": time(10)}]]]
     assert pl.Series(payload).to_list() == payload
-
-
-def test_struct_list_cat_8235() -> None:
-    df = pl.DataFrame(
-        {"values": [["a", "b", "c"]]}, schema={"values": pl.List(pl.Categorical)}
-    )
-    assert df.select(pl.struct("values")).to_dict(False) == {
-        "values": [{"values": ["a", "b", "c"]}]
-    }
 
 
 def test_struct_name_passed_in_agg_apply() -> None:
@@ -1007,127 +903,3 @@ def test_struct_name_passed_in_agg_apply() -> None:
             ]
         ],
     }
-
-
-def test_struct_lit_cast() -> None:
-    df = pl.DataFrame({"a": [1, 2, 3]})
-    schema = {"a": pl.Int64, "b": pl.List(pl.Int64)}
-
-    for lit in [pl.lit(None), pl.lit([[]])]:
-        s = df.select(pl.struct([pl.col("a"), lit.alias("b")], schema=schema))["a"]  # type: ignore[arg-type]
-        assert s.dtype == pl.Struct(
-            [pl.Field("a", pl.Int64), pl.Field("b", pl.List(pl.Int64))]
-        )
-        assert s.to_list() == [
-            {"a": 1, "b": None},
-            {"a": 2, "b": None},
-            {"a": 3, "b": None},
-        ]
-
-
-def test_struct_from_schema_only() -> None:
-    # we create a dataframe with default types
-    df = pl.DataFrame(
-        {
-            "str": ["a", "b", "c", "d", "e"],
-            "u8": [1, 2, 3, 4, 5],
-            "i32": [1, 2, 3, 4, 5],
-            "f64": [1, 2, 3, 4, 5],
-            "cat": ["a", "b", "c", "d", "e"],
-            "datetime": pl.Series(
-                [
-                    date(2023, 1, 1),
-                    date(2023, 1, 2),
-                    date(2023, 1, 3),
-                    date(2023, 1, 4),
-                    date(2023, 1, 5),
-                ]
-            ),
-            "bool": [1, 0, 1, 1, 0],
-            "list[u8]": [[1], [2], [3], [4], [5]],
-        }
-    )
-
-    # specify a schema with specific dtypes
-    s = df.select(
-        pl.struct(
-            schema={
-                "str": pl.Utf8,
-                "u8": pl.UInt8,
-                "i32": pl.Int32,
-                "f64": pl.Float64,
-                "cat": pl.Categorical,
-                "datetime": pl.Datetime("ms"),
-                "bool": pl.Boolean,
-                "list[u8]": pl.List(pl.UInt8),
-            }
-        ).alias("s")
-    )["s"]
-
-    # check dtypes
-    assert s.dtype == pl.Struct(
-        [
-            pl.Field("str", pl.Utf8),
-            pl.Field("u8", pl.UInt8),
-            pl.Field("i32", pl.Int32),
-            pl.Field("f64", pl.Float64),
-            pl.Field("cat", pl.Categorical),
-            pl.Field("datetime", pl.Datetime("ms")),
-            pl.Field("bool", pl.Boolean),
-            pl.Field("list[u8]", pl.List(pl.UInt8)),
-        ]
-    )
-
-    # check values
-    assert s.to_list() == [
-        {
-            "str": "a",
-            "u8": 1,
-            "i32": 1,
-            "f64": 1.0,
-            "cat": "a",
-            "datetime": datetime(2023, 1, 1, 0, 0),
-            "bool": True,
-            "list[u8]": [1],
-        },
-        {
-            "str": "b",
-            "u8": 2,
-            "i32": 2,
-            "f64": 2.0,
-            "cat": "b",
-            "datetime": datetime(2023, 1, 2, 0, 0),
-            "bool": False,
-            "list[u8]": [2],
-        },
-        {
-            "str": "c",
-            "u8": 3,
-            "i32": 3,
-            "f64": 3.0,
-            "cat": "c",
-            "datetime": datetime(2023, 1, 3, 0, 0),
-            "bool": True,
-            "list[u8]": [3],
-        },
-        {
-            "str": "d",
-            "u8": 4,
-            "i32": 4,
-            "f64": 4.0,
-            "cat": "d",
-            "datetime": datetime(2023, 1, 4, 0, 0),
-            "bool": True,
-            "list[u8]": [4],
-        },
-        {
-            "str": "e",
-            "u8": 5,
-            "i32": 5,
-            "f64": 5.0,
-            "cat": "e",
-            "datetime": datetime(2023, 1, 5, 0, 0),
-            "bool": False,
-            "list[u8]": [5],
-        },
-    ]
