+++ py-polars/tests/unit/test_empty.py
@@ -34,12 +34,6 @@ def test_empty_string_replace() -> None:
     assert s.str.replace("ab", "b").series_equal(s)
 
 
-def test_empty_duration() -> None:
-    s = pl.DataFrame([], {"days": pl.Int32}).select(pl.duration(days="days"))
-    assert s.dtypes == [pl.Duration("ns")]
-    assert s.shape == (0, 1)
-
-
 def test_empty_window_function() -> None:
     expr = (pl.col("VAL") / pl.col("VAL").sum()).over("KEY")
